Improve code style
